// app/(manager)/screens/types.ts
  export interface EmployerDetailsDisplayProps {
    id: string;
        name: string;
        email: string;
        phone: string;
        homeAddress: string;
        dateOfBirth: string;
        Language: string;
        identityVerificationStatus: string;
        backgroundCheckStatus: string;
        description: string;
        rating: number;
        assurance: number;
        urlToImage: any;
        imageStatus: string;
        IDCardStatus: string;
        PassportStatus: string;
        otherCertificateStatus: string;
  }
// ...
  export interface DisplayPersonalInfoProps {
    // key?: string;
    item: any;
}
// ...
export interface UserDataProps {
      id: string;
      name: string;
      email: string;
      phone: string;
      homeAddress: string;
      dateOfBirth: string;
      Language: string;
      identityVerificationStatus: string;
      backgroundCheckStatus: string;
      description: string;
      rating: string,
      assurance: string,
      urlToImage: any,
      imageStatus: string;
      IDCardStatus: string;
      PassportStatus: string;
      otherCertificateStatus: string;
  }